requirements.txt
blis==0.7.8
catalogue==2.0.8
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
cymem==2.0.6
es-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.4.0/es_core_news_lg-3.4.0-py3-none-any.whl
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.4.0/es_core_news_md-3.4.0-py3-none-any.whl
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.4.0/es_core_news_sm-3.4.0-py3-none-any.whl
Flask==2.2.2
fpdf==1.7.2
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.1
msgpack==1.0.4
murmurhash==1.0.8
numpy==1.23.2
packaging==21.3
pathy==0.6.2
Pillow==9.2.0
preshed==3.0.7
pydantic==1.9.2
pynvim==0.4.3
pyparsing==3.0.9
requests==2.28.1
smart-open==5.2.1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
thinc==8.1.0
tqdm==4.64.0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.12
wasabi==0.10.1
Werkzeug==2.2.2
